User Authentication & Login View Design

+++ app/Http/Controllers/ContactsController.php
@@ -3,34 +3,52 @@
 namespace App\Http\Controllers;
 
 use App\Contact;
+use App\Http\Resources\Contact as ContactResource;
 use Illuminate\Http\Request;
+use Symfony\Component\HttpFoundation\Response;
 
 class ContactsController extends Controller
 {
-    public function index(){
-    return request()->user()->contacts;
+    public function index()
+    {
+        $this->authorize('viewAny', Contact::class);
+        return ContactResource::collection(request()->user()->contacts);
     }
 
     public function store()
     {
-        Contact::create($this->validateData());
+        $this->authorize('create', Contact::class);
+        $contact = request()->user()->contacts()->create($this->validateData());
+        return (new ContactResource($contact))
+            ->response()
+            ->setStatusCode(Response::HTTP_ACCEPTED);
     }
 
     public function show(Contact $contact)
     {
-        return $contact;
+        $this->authorize('view', $contact);
+        return new ContactResource($contact);
     }
 
     public function update(Contact $contact)
     {
+        $this->authorize('update', $contact);
         $contact->update($this->validateData());
+
+        return (new ContactResource($contact))
+        ->response()
+        ->setStatusCode(Response::HTTP_OK);
     }
 
-    public function destroy(Contact $contact){
+    public function destroy(Contact $contact)
+    {
+        $this->authorize('delete', $contact);
         $contact->delete();
+        return response([], Response::HTTP_NO_CONTENT);
     }
 
-    private function validateData(){
+    private function validateData()
+    {
         return request()->validate([
             'name' => 'required',
             'email' => 'required|email',
